langchain/rag/requirements_rag.txt
langchain-core
langchain-community
langchain-postgres
langchain-openai

sqlalchemy>=2

psycopg[binary]
pgvector

fastapi
uvicorn[standard]
python-dotenv
pydantic

